Updates:
removed show bookings and delete buttons and changed to only Bookings
page and update asset

added delete button to the bookings page and then made it redirect back
to the same bookings page, however this meant that the mapping was the
delete mapping so if refreshed then error would show. Work around for
this is to add a if null statement to deletebooking method in asset
controller but this means calling a null booking id from DB will now
return null and not the exception as before.

Delete asset button moved to update asset page